requirements.txt
pandas>=2.0.0
numpy>=1.24.0
networkx>=3.0
matplotlib>=3.7.0
jellyfish>=1.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
openai>=1.12.0 